Move context-menu link capture into CmuxWebView+ContextMenuLinkCapture.swift
Keeps CmuxWebView.swift and BrowserConfigTests.swift inside the Swift
file length budget: the capture hook, link resolution helpers, and the
zoom-aware hit-test coordinate conversion move to a dedicated extension
file, and the regression test gets its own wired test file.
